Fix mem_warning_threshold type annotation (#627)
* Updated thresholds to float with 0.0 as the default value

* Update cpu_limit to float

* Add explicit type annotation for mypy

+++ api/resource_usage/src/jupyverse_resource_usage/__init__.py
@@ -31,7 +31,7 @@ class ResourceUsage(Router, ABC):
 
 class ResourceUsageConfig(Config):
     mem_limit: int = 0
-    mem_warning_threshold: int = 0
+    mem_warning_threshold: float = 0.0
     track_cpu_percent: bool = False
-    cpu_limit: int = 0
-    cpu_warning_threshold: int = 0
+    cpu_limit: float = 0.0
+    cpu_warning_threshold: float = 0.0
